Parcial 2 Tarea 2, Herencia y polimorfismo

+++ src/uaslp/ingenieria/labs/Shape.java
@@ -8,12 +8,11 @@ public class Shape {
         return "shape";
     }
 
-    // Tarea
     public int getSidesCount() {
         return 0;
     }
 
-    public int getPerimeter() {
+    public double getPerimeter() {
         return 0;
     }
